+++ src/DonorProfiles/Helpers.php
@@ -13,6 +13,13 @@ class Helpers {
 	 */
 	public static function getCurrentDonorId() {
 
+		if ( get_current_user_id() ) {
+			$donor = give()->donors->get_donor_by( 'user_id', get_current_user_id() );
+			if ( $donor ) {
+				return $donor->id;
+			}
+		}
+
 		if ( give()->email_access ) {
 			give()->email_access->init();
 			$useToken = give()->email_access->check_for_token();
@@ -23,8 +30,12 @@ class Helpers {
 			}
 		}
 
-		if ( get_current_user_id() ) {
-			$donor = give()->donors->get_donor_by( 'user_id', get_current_user_id() );
+		if (
+			false !== give()->session->get_session_expiration() ||
+			true === give_get_history_session()
+		) {
+			$email = give()->session->get( 'give_email' );
+			$donor = give()->donors->get_donor_by( 'email', $email );
 			if ( $donor ) {
 				return $donor->id;
 			}